Move generateInitialBands to dedicated driver
