+++ src/main/java/io/github/metal_pony/sudoku/drivers/Main.java
@@ -148,7 +148,7 @@ public class Main {
     put("countSolutions", Main::countSolutions);
     // --puzzle %s --threads %d --timeout %d
     put("solve", Main::solve);
-    put("generateBands", Main::generateInitialBands);
+    put("generateBands", GenerateInitialBands::generateInitialBands);
     // --level %d --grid %s --threads %d
     put("sieve", Main::createSieve);
     // --level %d --grid %s --threads %d
@@ -489,213 +489,12 @@ Commands:
   //   scanner.close();
   // }
 
-  private static long timeCpuExecution(Runnable runnable) {
+  static long timeCpuExecution(Runnable runnable) {
     ThreadMXBean bean = ManagementFactory.getThreadMXBean();
     long start = bean.getCurrentThreadCpuTime();
     runnable.run();
     long end = bean.getCurrentThreadCpuTime();
     return end - start;
-  }
-
-  static class Node2 {
-    Sudoku sudoku;
-    int index = -1;
-    int values = -1;
-    public Node2(Sudoku sudoku) {
-      this.sudoku = sudoku;
-      sudoku.reduce();
-      index = sudoku.pickEmptyCell(0, 27);
-      if (index != -1) {
-        values = sudoku.getCandidate(index);
-      }
-    }
-    public Node2 next() {
-      if (values <= 0) {
-          return null;
-      }
-      Sudoku s = new Sudoku(sudoku);
-      int d = Sudoku.CANDIDATES_ARR[values][0];
-      s.setDigit(index, d);
-      values &= ~(Sudoku.ENCODER[d]);
-      return new Node2(s);
-    }
-  }
-
-  public static void generateInitialBands(ArgsMap args) {
-    Set<String> fullBandSet = new HashSet<>();
-    final int N = Sudoku.DIGITS * 3;
-    long time = timeCpuExecution(() -> {
-      Sudoku root = new Sudoku("123456789--------");
-
-      // TODO Ensure that this script works properly without needing to call reset.
-      // root.resetCandidatesAndValidity();
-
-      Stack<Node2> q = new Stack<>();
-      q.push(new Node2(root));
-
-      while (!q.isEmpty()) {
-        Node2 top = q.peek();
-        Node2 next = top.next();
-        if (next == null) {
-          boolean hasEmptyInBand = top.sudoku.pickEmptyCell(0, N) >= 0;
-          if (!hasEmptyInBand) {
-            String bandStr = top.sudoku.toString().substring(0, N);
-            if (fullBandSet.add(bandStr)) {
-              if (verbose) {
-                System.out.println(bandStr);
-              }
-            }
-          }
-          q.pop();
-        } else {
-          q.push(next);
-        }
-      }
-    });
-
-    if (verbose) {
-      System.out.printf(
-        " -- found %d initial bands in %s ms --\n",
-        fullBandSet.size(),
-        TimeUnit.NANOSECONDS.toMillis(time)
-      );
-      System.out.println("Reducing bands...");
-    }
-
-    long startTime = System.currentTimeMillis();
-    Set<String> reducedBandSet = reduceFullBandSet(fullBandSet);
-    long endTime = System.currentTimeMillis();
-    if (verbose) {
-      System.out.printf(" -- reduced bands to %d in %d ms --\n", reducedBandSet.size(), (endTime - startTime));
-    }
-    reducedBandSet.forEach(System.out::println);
-  }
-
-  public static Set<String> reduceFullBandSet(Set<String> fullBandSet) {
-    // TODO Reduce fullBandSet by discovering and removing transforms
-    // For each BAND:
-    //  new queue, new hashset<string> to track seen elements, add BAND
-    //  while queue not empty:
-    //    b = poll
-    //    // always normalize after transform, before adding to queue
-    //    add unseen block permutations to queue,
-    //    add unseen row permutations to queue,
-    //    add unseen column permutations to queue,
-    //    band -> config -> search for UAs(level 2? 3?) -> when found, if (bandMask & ua) == ua -> if unseen, add to queue
-
-    List<String> allBands = new ArrayList<>(fullBandSet);
-    HashSet<String> reducedBands = new HashSet<>();
-    final int N = Sudoku.DIGITS * 3;
-
-    while (!allBands.isEmpty()) {
-      String band = allBands.remove(allBands.size() - 1);
-      // String bandPuzzleStr = band + "0".repeat(Sudoku.SPACES - band.length());
-      // Sudoku bandPuzzle = new Sudoku(bandPuzzleStr);
-      HashSet<String> seen = new HashSet<>();
-      Queue<String> q = new LinkedList<>();
-      seen.add(band);
-      q.offer(band);
-      reducedBands.add(band);
-
-      // TODO NOT FEASIBLE TO COUNT SOLUTIONS
-      // int rootCount = countSolutions(new Sudoku(band + "0".repeat(Sudoku.SPACES - band.length())));
-      // if (verbose) {
-      //   System.out.printf("Transforming band %s, all transforms should have %d solutions:\n", band, rootCount);
-      // }
-
-      while (!q.isEmpty()) {
-        String bStr = q.poll() + "0".repeat(Sudoku.SPACES - band.length());
-
-        // Transforms
-        Sudoku[] transforms = new Sudoku[] {
-          new Sudoku(bStr).swapStacks(1, 2),
-          new Sudoku(bStr).swapStacks(0, 1),
-          new Sudoku(bStr).swapStacks(0, 1).swapStacks(1, 2),
-          new Sudoku(bStr).swapStacks(0, 2).swapStacks(1, 2),
-          new Sudoku(bStr).swapStacks(0, 2),
-
-          new Sudoku(bStr).swapBandRows(0, 1, 2),
-          new Sudoku(bStr).swapBandRows(0, 0, 1),
-          new Sudoku(bStr).swapBandRows(0, 0, 1).swapBandRows(0, 1, 2),
-          new Sudoku(bStr).swapBandRows(0, 0, 2).swapBandRows(0, 1, 2),
-          new Sudoku(bStr).swapBandRows(0, 0, 2),
-
-          new Sudoku(bStr).swapStackCols(0, 1, 2),
-          new Sudoku(bStr).swapStackCols(0, 0, 1),
-          new Sudoku(bStr).swapStackCols(0, 0, 1).swapStackCols(0, 1, 2),
-          new Sudoku(bStr).swapStackCols(0, 0, 2).swapStackCols(0, 1, 2),
-          new Sudoku(bStr).swapStackCols(0, 0, 2),
-
-          new Sudoku(bStr).swapStackCols(1, 1, 2),
-          new Sudoku(bStr).swapStackCols(1, 0, 1),
-          new Sudoku(bStr).swapStackCols(1, 0, 1).swapStackCols(1, 1, 2),
-          new Sudoku(bStr).swapStackCols(1, 0, 2).swapStackCols(1, 1, 2),
-          new Sudoku(bStr).swapStackCols(1, 0, 2),
-
-          new Sudoku(bStr).swapStackCols(2, 1, 2),
-          new Sudoku(bStr).swapStackCols(2, 0, 1),
-          new Sudoku(bStr).swapStackCols(2, 0, 1).swapStackCols(2, 1, 2),
-          new Sudoku(bStr).swapStackCols(2, 0, 2).swapStackCols(2, 1, 2),
-          new Sudoku(bStr).swapStackCols(2, 0, 2)
-        };
-
-        for (Sudoku t : transforms) {
-          String tStr = t.normalize().toString().substring(0, N);
-          if (!seen.contains(tStr)) {
-            seen.add(tStr);
-            q.offer(tStr);
-
-            // TODO NOT FEASIBLE TO COUNT SOLUTIONS
-            // int count = countSolutions(new Sudoku(tStr));
-            // if (verbose) {
-            //   System.out.printf("%s [%d] %s\n", (rootCount == count) ? "  " : "🚨", count, tStr);
-            // }
-          }
-        }
-
-        // TODO Additional symmetries can be found by locating UAs within the band
-
-        // AtomicReference<Sudoku> atomicConfig = new AtomicReference<>();
-        // bandPuzzle.searchForSolutions3(solution -> {
-        //   atomicConfig.set(solution);
-        //   return false;
-        // });
-        // Sudoku c = atomicConfig.get();
-        // SudokuSieve sieve = new SudokuSieve(c.getBoard());
-        // BigInteger bandMask = new BigInteger("1".repeat(N) + "0".repeat(Sudoku.SPACES - N), 2);
-        // for (int r = Sudoku.DIGIT_COMBOS_MAP[2].length - 1; r >= 0; r--) {
-        //   BigInteger pMask = c.maskForDigits(Sudoku.DIGIT_COMBOS_MAP[2][r]);
-        //   sieve.addFromFilter(pMask, (solution) -> {
-        //     // TODO item may need to be inverted
-        //     BigInteger item = c.diff2(solution);
-        //     if (item.equals(item.and(bandMask))) {
-        //       String tStr = solution.normalize().toString().substring(0, N);
-        //       if (!seen.contains(tStr)) {
-        //         seen.add(tStr);
-        //         q.offer(tStr);
-        //       }
-        //     }
-        //   });
-        // }
-      }
-
-      int sizeBefore = allBands.size();
-      allBands.removeAll(seen);
-      int sizeAfter = allBands.size();
-      if (verbose) {
-        System.out.printf("Removed %d permuted bands, (%d remaining).\n", sizeBefore - sizeAfter, allBands.size());
-      }
-    }
-
-    if (verbose) {
-      System.out.printf(
-        "Done.\nRemoved %d permuted bands in total.\nReduced band set size: %d.\n",
-        fullBandSet.size() - reducedBands.size(),
-        reducedBands.size()
-      );
-    }
-
-    return reducedBands;
   }
 
   /**
